backend/requirements.txt
pymupdf==1.25.3
python-pptx==1.0.2
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.18
aiofiles==24.1.0
aiosqlite==0.20.0
httpx==0.27.0
python-dotenv==1.0.0
Pillow
numpy
